writable2

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -112,7 +112,7 @@ public class PairsPMI  extends Configured implements Tool {
 
 
   private static final class MyMapper2 extends Mapper<LongWritable, Text, PairOfStrings, FloatWritable> {
-    private static final IntWritable ONE = new IntWritable(1);
+    private static final FloatWritable ONE = new FloatWritable(1);
     private static final PairOfStrings Pair = new PairOfStrings();
 
     @Override
